Fix variable serialization issue
For some reason including dummy_index in the serialization and
deserialization process breaks the fact that two references point
to the same object once they're deserialized. To workaround this,
I have removed the dummy_index from the serialization. However,
this means that a component C is serialized and then deserialized
into component D, the parameters in C will not be equal to the
parameters in D because their dummy indices will differ.